pyproject.toml
[tool.poetry]
name = "flask-admin-skel"
version = "0.1.0"
description = "Flask Admin Skeleton"
authors = ["Mario Hernandez <mariofix@proton.me>"]
license = "MIT"
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
flask-admin = "^1.6.1"
flask-migrate = "^4.0.4"
flask = "^2.2.3"
python-dotenv = "^1.0.0"
flask-babel = "^3.0.1"
sqlalchemy-repr = "^0.1.0"
thankyou = "^0.0.1"
pymysql = "^1.0.2"
bcrypt = "^4.0.1"
python-slugify = "^8.0.1"
python-multipart = "^0.0.6"
requests = "^2.28.2"
httpx = "^0.23.3"
python-dateutil = "^2.8.2"
email-validator = "^1.3.1"
sqlalchemy-utils = "^0.40.0"
flask-http-middleware = "^0.2.1"
flask-security-too = {extras = ["babel", "common", "sqla"], version = "^5.1.1"}


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pytest = "^7.2.2"
coverage = "^7.2.1"
flask-debugtoolbar = {git = "https://github.com/pallets-eco/flask-debugtoolbar.git", rev = "master"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
